backend/python/requirements.txt
flask==3.0.0
flask-cors==4.0.0
pandas==2.1.4
numpy==1.26.2
yfinance==0.2.33
requests==2.31.0
python-dotenv==1.0.0
